+++ src/api/config.ts
@@ -1,7 +1,7 @@
 export const API_SERVER_VERSION='0.1.8';
 export const API_SERVER_VERSION_TAG='latest';
 
-let baseUrl = 'http://localhost:4000';
+let baseUrl = 'http://localhost/ocdb-server';
 
 if (process.env.NODE_ENV === 'production') {
     baseUrl = "https://ocdb.eumetsat.int";